Organise code into suitable packages
