+++ src/main/java/duke/command/AddCommand.java
@@ -1,5 +1,6 @@
-package duke;
+package duke.command;
 
+import duke.Ui;
 import duke.storage.Storage;
 import duke.task.Task;
 import duke.task.TaskFactory;
